Renammed metaSwitch action to metaToggle
